services/auth/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
cryptography==41.0.7
python-dotenv==1.0.0
